gpt_telegram/requirements.txt
openai==0.27.2
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
requests==2.25.1
